Create Room Joining Mechanics
After creating the endpoint to create a room, I started on the
functions that will handle joining a created room or existing room.
However, it seems like the members of the rooms will have to be stored
somewhere , which has been commented in as a TODO.

joinRoom essentially has the socket log what it will do on the server,
which is joining socket with Player X to ROom A. and then the player
id is broadcasted to all other sockets within the room for it to be
updated. However a flaw that seems obvious is that new sockets that
join will have no knowledged of the previous sockets that are
currently in the room which can be said to be the same as any
reconnecting socket.

leaveRoom performs a similar function, however with the difference being
that it has the task of removing the socket from the room.