Update Room Joining Mechanics I
Added db transactions which would save the current listing of players
in the room. Updated the Schema again to reflect the changes to the
DB. Also now when a player is added or removed from the room, there is
a call to the DB to return a current listing of all players to ensure
that new sockets will be able to have a current listing of players.

TODO: Test the room creation as well as joining and leaving a room.