Setup Game Mechanics
Create basic Game schema which includes the roomid that it tied to as
well as created and updated date.

Created the basic gamehandler file for all the socket methods and
events to be handled by the server

I honestly have no idea how to code this game mechanic in code as yet.
I guess I should start with how the game works then break it up piece
by piece but this is the commit for today