Create Game Mechanics I
created a class Game in Game.js to encapsulate how the game will be
played as a sounding board. I have created the basic functions for
things I think I will need such as how many rounds the game will go to
, currentPlayer, Turns in the Game, currentRounds. I think I may need
more , need to visualize the game or see it happen more.

Thats the commmit for today =)