Create Player Class
I realized that some of the code regarding the players may be need to
be in its own class. So I have moved the storage of a player status to
the class. Also the players will have to record a username as well as
a balance that they will be making bets from so I have added that to
the class as well as setter and getters matching those values. Also
added, some functions to check the status of the players as well as
making bets. More research to be conducted to fledge out game play.

Thats the commit for today =)