Update Player.js and Game.js
Since I had created some player functions yesterday which dealt with
handling player state, I needed to update some of the game functions
to utilise some of those functions, also added isPlayerEligibleToBet
and addPlayerBetToPot, to new functions for hadnling updating the
gamePot variablle in the Game class. Side note, learnt some more of
how betting is done in Poker, which would help me flesh out this idea
more, tho more research to be done.

That is the commit for today =)