Create Deck.js and Card.js, update Player.js
Create the code representation of the Deck and Cards in the Backend. A
deck is comprised of two arrays, one to model the hidden cards that will
be show to the player and one to model the visible cards that will be shown to
will be shown to other players and not to the player. A card stores the
basic number value it represents and the can be used or not which help
to reflect state in the frontend. Player Class was modified to store a
deck for each user. Functions were created to help manage state changes
to the deck. Found my earlier notes on the game idea so it will be
helping to refine what I think of implementing as the MVP.

Thats the commit for today =)