Create Player Model, update Game Model + game mech
After working on the classes to model the game logic, I realize that
this solution requires for the instances to persist so it would seem
to be better to save information in the db rather than work through
the classes to keep track of the game at this moment. So A Player
Model was created to reflect the current state of players, a game will
be created upon the game:create event sent from the client that owns a
room or is the admin. Created the stub for the startGame functions.
Also the roomId will be equivalent to the gameid so that saves the
decision of saving two different ids.

Thats the commit for today =)