create setup & start game mech.
Created the socket handler server side for setting up the game behind
the scenes, added functions which intialized the players with there
decks as well as starting balances, as wellas perform various db
querys and updates atleast on th player collection. Added minimal
error logic in the execptions of current and past fucntions in the
handlers, feel like Im starting to get a flow of how the game might
perform. Still more to be research and done. Might work on some
frontend to test what I have been working on for the past two weeks to
ensure it is working well.

Thats the commit for today =)