Update Player Action Functions I
Added the pot update after the player makes an opening bet via the bet
player action. Also created some functions that support this action as
well as query the DB. Modified the Game model to house a gamePot
column so a value can be stored for the gamePot.

Thats the commit for today =)